+++ T850/Framework/src/video/GLRT.cpp
@@ -244,7 +244,7 @@ namespace t800 {
   void GLRT::DestroyAPIRT() {
     GLuint FBO = vFrameBuffers[0];
     glDeleteFramebuffers(1, &FBO);
-    for (int i = 0; i < vColorTextures.size(); i++) {
+    for (size_t i = 0; i < vColorTextures.size(); i++) {
       vColorTextures[i]->release();
     }
     pDepthTexture->release();
